Cell implementation

+++ Java Project/Grass.java
@@ -15,7 +15,7 @@ public class Grass extends Vegetation {
     }
 
     @Override
-    public float calculateSpreadHeat(float fireIntensity) {
+    public float calculateSpreadHeat(int fireIntensity) {
         // Calls base calculation from Vegetation and factors in density
         return calculateSpreadHeat(fireIntensity) + this.density;
     }
